refactor: adjust import and extension function
