+++ backend/framework/src/test/kotlin/br/com/zup/beagle/serialization/jackson/BeagleBuilderSerializerTest.kt
@@ -16,6 +16,7 @@
 
 package br.com.zup.beagle.serialization.jackson
 
+import br.com.zup.beagle.widget.layout.ComposeComponent
 import br.com.zup.beagle.widget.layout.Screen
 import br.com.zup.beagle.widget.layout.ScreenBuilder
 import br.com.zup.beagle.widget.ui.Text
