+++ backend/framework/src/test/kotlin/br/com/zup/beagle/serialization/jackson/BeagleSerializerModifierTest.kt
@@ -20,6 +20,7 @@ import br.com.zup.beagle.action.Action
 import br.com.zup.beagle.action.Navigate
 import br.com.zup.beagle.core.ServerDrivenComponent
 import br.com.zup.beagle.widget.Widget
+import br.com.zup.beagle.widget.layout.ComposeComponent
 import br.com.zup.beagle.widget.layout.Screen
 import br.com.zup.beagle.widget.layout.ScreenBuilder
 import com.fasterxml.jackson.databind.BeanDescription
