+++ schema/kotlin-widgets/src/main/kotlin/br/com/zup/beagle/widget/ui/ListView.kt
@@ -56,21 +56,3 @@ data class ListView(
 ) : ServerDrivenComponent {
     companion object
 }
-
-
-
-/**
- * Represent a view in list
- *
- */
-
-typealias RowBuilder = (index: Int) -> Widget
-
-fun ListView.Companion.dynamic(
-    size: Int,
-    direction: ListDirection = ListDirection.VERTICAL,
-    rowBuilder: RowBuilder
-) = ListView(
-    rows = (0 until size).map(rowBuilder),
-    direction = direction
-)
